requirements.txt
torch
torchvision
pillow
numpy
einops
thop
timm